src/eDoxa.Identity.Api/Areas/Identity/Controllers/AddressController.cs
// Filename: AddressController.cs
// Date Created: 2019-08-09
// 
// ================================================
// Copyright © 2019, eDoxa. All rights reserved.

using System.Threading.Tasks;

using AutoMapper;

using eDoxa.Identity.Api.Areas.Identity.Responses;
using eDoxa.Identity.Api.Areas.Identity.Services;

using IdentityServer4.AccessTokenValidation;

using Microsoft.AspNetCore.Authorization;
using Microsoft.AspNetCore.Mvc;

namespace eDoxa.Identity.Api.Areas.Identity.Controllers
{
    [ApiController]
    [ApiVersion("1.0")]
    [Produces("application/json")]
    [Route("api/address")]
    [ApiExplorerSettings(GroupName = "Address")]
    [Authorize(AuthenticationSchemes = IdentityServerAuthenticationDefaults.AuthenticationScheme)]
    public class AddressController : ControllerBase
    {
        private readonly ICustomUserManager _userManager;
        private readonly IMapper _mapper;

        public AddressController(ICustomUserManager userManager, IMapper mapper)
        {
            _userManager = userManager;
            _mapper = mapper;
        }

        /// <summary>
        ///     Find user's address.
        /// </summary>
        [HttpGet]
        public async Task<IActionResult> GetAsync()
        {
            var user = await _userManager.GetUserAsync(User);

            var address = await _userManager.GetAddressAsync(user);

            if (address == null)
            {
                return this.NoContent();
            }

            return this.Ok(_mapper.Map<AddressResponse>(address));
        }
    }
}
